requirements.txt
# Core dependencies
playwright==1.49.1
playwright-stealth==1.0.6
pydantic==2.10.5
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.3.0
pymongo==4.6.1
motor==3.3.2
dnspython==2.4.2

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
py2neo>=2021.2.3

# CLI
click>=8.0.0
pyyaml>=6.0
